Rename : useSetting hook 이름변경
